reuseable code moved to util function
